app/backend/book-store/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
pymongo==4.8.0
requests==2.32.3
timeout-decorator==0.5.0
boto3==1.38.28
botocore==1.38.28
python-dotenv==1.0.1
python-jose==3.3.0
wraps==0.15.0
flasgger==0.9.7.1
certifi>=2024.0.0
aws-xray-sdk==2.12.0